requirements.txt

Flask==2.3.3
Flask-SQLAlchemy==3.0.3
gunicorn==21.2.0
Pillow==10.2.0
psd-tools==1.12.1
reportlab==4.0.8
